Made progression pie and radius marker working inside of trouble brewing on the Swarm NPC.
KNOWN ISSUE:
If the local player enters the area of ANY swarm npc, that swarm npc will turn red, eventhough it can despawn if another player spawned it. fix this? same with rendering swarms that arent spawned by the player. since those dont reaallly matter? or detect the player that spawned it within that area!?